category bar added
