+++ src/components/restaurantDetails/MenuList/index.jsx
@@ -1,41 +1,20 @@
 import React, { useRef } from 'react';
-import { StyledMenuWrapper, MenuProps } from './menuList.styled';
-import { Grid, Typography, InputLabel, MenuItem, FormControl, Select, useTheme, OutlinedInput } from '@mui/material';
+import { StyledMenuWrapper, StyledOptionBar, StyledOptionItem } from './menuList.styled';
+import { Grid, Typography, Stack, Divider } from '@mui/material';
 import MenuCard from '../MenuCard/index';
 import MenuModal from '../MenuModal/index';
 
-function getStyles(name, menuName, theme) {
-  return {
-    fontWeight: menuName.indexOf(name) === -1 ? theme.typography.fontWeightRegular : theme.typography.fontWeightBold,
-  };
-}
-
 export default function MenuList(props) {
-  const theme = useTheme();
   const [open, setOpen] = React.useState(false);
   const [menuName, setMenuName] = React.useState('');
   const [menuDesc, setMenuDesc] = React.useState('');
   const [menuIngri, setMenuIngri] = React.useState([]);
   const [imgUrl, setImgUrl] = React.useState('');
   const [price, setPrice] = React.useState('');
-
   let refs = [useRef(null), useRef(null)];
 
   const handleClose = () => {
     setOpen(false);
-  };
-  const handleChange = (event) => {
-    const {
-      target: { value },
-    } = event;
-    setMenuName(value);
-    refs.forEach((el, index) => {
-      if (el.current.textContent === value) {
-        // scrolling doesn't work when there's optional parameter
-        // tried enabling chrome://flags/#smooth-scrolling
-        refs[index].current.scrollIntoView({ behaviour: 'smooth' });
-      }
-    });
   };
 
   const handleCardClick = (menuName, menuDesc, menuIngri, imgUrl, price) => {
@@ -46,7 +25,13 @@ export default function MenuList(props) {
     setPrice(price);
     setOpen(true);
   };
-
+  const handleCateogryClick = (value) => {
+    refs.forEach((el, index) => {
+      if (el.current.textContent === value) {
+        refs[index].current.scrollIntoView({ block: 'start', behaviour: 'smooth' });
+      }
+    });
+  };
   return (
     <>
       <MenuModal
@@ -58,25 +43,16 @@ export default function MenuList(props) {
         menuPrice={price}
         handleClose={handleClose}
       />
-      {/* <Grid container top="0" position="sticky" direction="row-reverse">
-        <FormControl sx={{ m: 1, width: 300 }}>
-          <InputLabel>Greasy's Burgers</InputLabel>
-          <Select
-            id="menu-category"
-            value={menuName}
-            onChange={handleChange}
-            input={<OutlinedInput label="Name" />}
-            MenuProps={MenuProps}
-            defaultValue=""
-          >
-            {props.menuCategory.map((menu) => (
-              <MenuItem key={menu} value={menu} style={getStyles(menu, menuName, theme)}>
-                {menu}
-              </MenuItem>
-            ))}
-          </Select>
-        </FormControl>
-      </Grid> */}
+      <StyledOptionBar>
+        <Stack direction="row" spacing={2}>
+          {props.menu.map((item, idx) => (
+            <StyledOptionItem key={idx} onClick={() => handleCateogryClick(item.category)}>
+              {item.category}
+            </StyledOptionItem>
+          ))}
+        </Stack>
+      </StyledOptionBar>
+      <Divider />
 
       <StyledMenuWrapper>
         <Grid container justifyContent="center" alignItems="center" spacing={2}>
