+++ src/components/restaurantDetails/MenuList/menuList.styled.js
@@ -1,5 +1,5 @@
 import { styled } from '@mui/styles';
-import { Box } from '@mui/material';
+import { Box, Grid, Button } from '@mui/material';
 
 const ITEM_HEIGHT = 48;
 const ITEM_PADDING_TOP = 8;
@@ -16,3 +16,26 @@ export const MenuProps = {
     },
   },
 };
+
+export const StyledOptionBar = styled(Grid)(({ theme }) => ({
+  backgroundColor: 'white',
+  flex: 'wrap',
+  display: 'flex',
+  padding: '16px',
+  position: 'sticky',
+  top: '0',
+  zIndex: 5,
+}));
+
+export const StyledOptionItem = styled(Button)(({ theme, selected }) => ({
+  // need to get theme primary color from props
+  backgroundColor: selected ? '#22b39a' : 'white',
+  borderRadius: selected ? '16px' : '',
+  color: selected ? 'white' : '#22b39a',
+  justifyContent: 'center',
+  padding: '2px 16px',
+  textAlign: 'center',
+  lineHeight: '24px',
+
+  '&:hover, &.Mui-focusVisible': { color: 'grey' },
+}));
